Provide exempt_domains example
When I was setting this up, I noticed that using the https protocol would cause the package to throw an exception. After a bit of tinkering, I realized that the domain needs to be completely stripped (no protocol) in order for the central app to recognize it. This PR just adds a simple example to the docs for this.