Provide exempt_domains examples (#44)
* Provide exempt_domains example

When I was setting this up, I noticed that using the https protocol would cause the package to throw an exception. After a bit of tinkering, I realized that the domain needs to be completely stripped (no protocol) in order for the central app to recognize it. This PR just adds a simple example to the docs for this.

* Update central-app.blade.md

* Update central-app.blade.md

* Update docs/source/v2/central-app.blade.md

Co-Authored-By: Samuel Štancl <samuel.stancl@gmail.com>

* Update docs/source/v2/configuration.blade.md

* ` instead of ```

* ` instead of ```

Co-authored-by: Samuel Štancl <samuel.stancl@gmail.com>