Subdomain/ Domain Clarification in Tenant Identification Docs (#56)
I believe this paragraph has the information switched. A string in the `domain` column that contains a **dot** would represent a domain/hostname and a string in this column without a dot, would represent a subdomain. The current documentation has this flipped. 

It had me second guessing myself. Unless I am misunderstanding something.

Maybe I am misunderstanding things, and if so, then disregard this PR.