Data column explanation + updating tenants
I've clarified how the data column works along with how to update a tenant's custom column and an attribute in its data column. I've also cleared up how the getCustomColumns() function works (When I first used the package I thought it meant to define columns that will be used in the data column).